split elc preprocessing into generation, grid, storage
